redux関連を分離したった
