+++ src/js/redux/actions.js
@@ -1,0 +1,30 @@
+export const PLAY = "play";
+export const PLAY_ALL = "play_all";
+export const PLAY_NEXT = "play_next";
+export const CLEAR = "clear";
+
+export const play = (track) => {
+    return {
+        type: PLAY,
+        track: track
+    };
+};
+
+export const playAll = (tracks) => {
+    return {
+        type: PLAY_ALL,
+        tracks: tracks
+    };
+};
+
+export const playNext = () => {
+    return {
+        type: PLAY_NEXT
+    };
+};
+
+export const clear = () => {
+    return {
+        type: CLEAR
+    };
+};
